feat(sandbox): unify remote execution into tabbed hub — Unsandbox, Tailscale, Kubernetes
Rebuild SandboxDashboard as a three-tab Remote Execution page.
Sidebar label: Sandbox → Remote. RuntimeConnectionsPane removed from
Settings — config now lives in the Remote page alongside execution.
