+++ apps/desktop/src/layout/sections.tsx
@@ -59,7 +59,7 @@ export const sidebarItems: SidebarItem[] = [
   },
   {
     id: 'SANDBOX',
-    label: 'Sandbox',
+    label: 'Remote',
     description: 'Remote code execution',
     icon: SandboxIcon,
   },
@@ -127,7 +127,7 @@ const sectionMeta: Record<SectionID, { label: string; title: string }> = {
   PROJECTS: { label: 'Workspace', title: 'Projects' },
   AGENTS: { label: 'Compute', title: 'Agents' },
   WAREHOUSE: { label: 'Usage', title: 'Usage' },
-  SANDBOX: { label: 'Compute', title: 'Sandbox' },
+  SANDBOX: { label: 'Compute', title: 'Remote Execution' },
   SETTINGS: { label: 'System', title: 'Settings' },
   DOCS: { label: 'Knowledge', title: 'Documentation' },
   CONSOLE: { label: 'Workspace', title: 'Development' },
